feat(sfcc): split SFCC in Flaps channel and Slats channel
