+++ fbw-a32nx/src/wasm/systems/a320_systems/src/hydraulic/mod.rs
@@ -78,8 +78,10 @@ use systems::{
     },
 };
 
-mod flaps_computer;
-use flaps_computer::SlatFlapComplex;
+mod flaps_channel;
+mod sfcc;
+mod slats_channel;
+use sfcc::SlatFlapComplex;
 
 #[cfg(test)]
 use systems::hydraulic::PressureSwitchState;
