+++ fbw-a32nx/src/wasm/systems/a320_systems/src/hydraulic/sfcc.rs
@@ -1,7 +1,6 @@
 use crate::systems::shared::arinc429::{Arinc429Word, SignStatus};
 use systems::accept_iterable;
-use systems::hydraulic::command_sensor_unit::{CSUMonitor, CSU};
-use systems::shared::{AdirsMeasurementOutputs, PositionPickoffUnit};
+use systems::shared::PositionPickoffUnit;
 
 use systems::simulation::{
     InitContext, SimulationElement, SimulationElementVisitor, SimulatorWriter, UpdateContext,
@@ -9,7 +8,10 @@ use systems::simulation::{
 };
 
 use std::panic;
-use uom::si::{angle::degree, f64::*, length::foot, velocity::knot};
+use uom::si::{angle::degree, f64::*};
+
+use super::flaps_channel::FlapsChannel;
+use super::slats_channel::SlatsChannel;
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 enum FlapsConf {
@@ -17,274 +19,77 @@ enum FlapsConf {
     Conf1,
     Conf1F,
     Conf2,
-    Conf2S,
     Conf3,
     ConfFull,
 }
 
 struct SlatFlapControlComputer {
-    flaps_conf_index_id: VariableIdentifier,
+    config_index_id: VariableIdentifier,
     slat_flap_system_status_word_id: VariableIdentifier,
     slat_flap_actual_position_word_id: VariableIdentifier,
-    slat_actual_position_word_id: VariableIdentifier,
-    flap_actual_position_word_id: VariableIdentifier,
 
-    flaps_demanded_angle: Angle,
-    slats_demanded_angle: Angle,
-    flaps_feedback_angle: Angle,
-    slats_feedback_angle: Angle,
-    flaps_handle_position: CSU,
-    flaps_conf: FlapsConf,
-    flap_load_relief_active: bool,
-    cruise_baulk_active: bool,
-    alpha_speed_lock_active: bool,
+    flaps_channel: FlapsChannel,
+    slats_channel: SlatsChannel,
 
-    csu_monitor: CSUMonitor,
+    config: FlapsConf,
 }
 
 impl SlatFlapControlComputer {
     const EQUAL_ANGLE_DELTA_DEGREE: f64 = 0.177;
-    const HANDLE_ONE_CONF_AIRSPEED_THRESHOLD_KNOTS: f64 = 205.;
-    const CONF1F_TO_CONF1_AIRSPEED_THRESHOLD_KNOTS: f64 = 212.;
-    const CRUISE_BAULK_AIRSPEED_THRESHOLD_KNOTS: f64 = 265.5;
-    const CRUISE_BAULK_ALTITUDE_THRESHOLD_FEET: f64 = 22000.;
-    const ALPHA_SPEED_LOCK_IN_AIRSPEED_THRESHOLD_KNOTS: f64 = 155.;
-    const ALPHA_SPEED_LOCK_OUT_AIRSPEED_THRESHOLD_KNOTS: f64 = 161.;
-    const ALPHA_SPEED_LOCK_IN_AOA_THRESHOLD_DEGREES: f64 = 9.5;
-    const ALPHA_SPEED_LOCK_OUT_AOA_THRESHOLD_DEGREES: f64 = 9.2;
-
-    const FLRS_CONFFULL_TO_CONF3_AIRSPEED_THRESHOLD_KNOTS: f64 = 184.5;
-    const FLRS_CONF3_TO_CONF2S_AIRSPEED_THRESHOLD_KNOTS: f64 = 198.5;
-    const FLRS_CONF2_TO_CONF1F_AIRSPEED_THRESHOLD_KNOTS: f64 = 222.5;
-    const FLRS_CONF1F_TO_CONF2_AIRSPEED_THRESHOLD_KNOTS: f64 = 217.5;
-    const FLRS_CONF2S_TO_CONF3_AIRSPEED_THRESHOLD_KNOTS: f64 = 193.5;
-    const FLRS_CONF3_TO_CONFFULL_AIRSPEED_THRESHOLD_KNOTS: f64 = 179.5;
 
     fn new(context: &mut InitContext, num: u8) -> Self {
         Self {
-            flaps_conf_index_id: context.get_identifier("FLAPS_CONF_INDEX".to_owned()),
+            config_index_id: context.get_identifier("FLAPS_CONF_INDEX".to_owned()),
             slat_flap_system_status_word_id: context
                 .get_identifier(format!("SFCC_{num}_SLAT_FLAP_SYSTEM_STATUS_WORD")),
             slat_flap_actual_position_word_id: context
                 .get_identifier(format!("SFCC_{num}_SLAT_FLAP_ACTUAL_POSITION_WORD")),
-            slat_actual_position_word_id: context
-                .get_identifier(format!("SFCC_{num}_SLAT_ACTUAL_POSITION_WORD")),
-            flap_actual_position_word_id: context
-                .get_identifier(format!("SFCC_{num}_FLAP_ACTUAL_POSITION_WORD")),
 
-            flaps_demanded_angle: Angle::new::<degree>(0.),
-            slats_demanded_angle: Angle::new::<degree>(0.),
-            flaps_feedback_angle: Angle::new::<degree>(0.),
-            slats_feedback_angle: Angle::new::<degree>(0.),
-            flaps_handle_position: CSU::Conf0,
-            flaps_conf: FlapsConf::Conf0,
-            flap_load_relief_active: false,
-            cruise_baulk_active: false,
-            alpha_speed_lock_active: false,
-
-            csu_monitor: CSUMonitor::new(context),
+            flaps_channel: FlapsChannel::new(context, num),
+            slats_channel: SlatsChannel::new(context, num),
+            config: FlapsConf::Conf0,
         }
-    }
-
-    // Returns a flap demanded angle in FPPU reference degree (feedback sensor)
-    // Interpolated from A320 FPPU references, assuming we're using the A320 FPPU
-    fn demanded_flaps_fppu_angle_from_conf(flap_conf: FlapsConf) -> Angle {
-        match flap_conf {
-            FlapsConf::Conf0 => Angle::new::<degree>(0.),
-            FlapsConf::Conf1 => Angle::new::<degree>(0.),
-            FlapsConf::Conf1F => Angle::new::<degree>(108.28),
-            FlapsConf::Conf2 => Angle::new::<degree>(154.65),
-            FlapsConf::Conf2S => Angle::new::<degree>(154.65),
-            FlapsConf::Conf3 => Angle::new::<degree>(194.03),
-            FlapsConf::ConfFull => Angle::new::<degree>(218.91),
-        }
-    }
-
-    // Returns a slat demanded angle in FPPU reference degree (feedback sensor)
-    // Interpolated from A320 FPPU references, assuming we're using the A320 FPPU
-    fn demanded_slats_fppu_angle_from_conf(flap_conf: FlapsConf) -> Angle {
-        match flap_conf {
-            FlapsConf::Conf0 => Angle::new::<degree>(0.),
-            FlapsConf::Conf1 => Angle::new::<degree>(247.27),
-            FlapsConf::Conf1F => Angle::new::<degree>(247.27),
-            FlapsConf::Conf2 => Angle::new::<degree>(247.27),
-            FlapsConf::Conf2S => Angle::new::<degree>(284.65),
-            FlapsConf::Conf3 => Angle::new::<degree>(284.65),
-            FlapsConf::ConfFull => Angle::new::<degree>(284.65),
-        }
-    }
-
-    fn generate_configuration(
-        &self,
-        context: &UpdateContext,
-        adirs: &impl AdirsMeasurementOutputs,
-    ) -> FlapsConf {
-        // Ignored `CSU::OutOfDetent` and `CSU::Fault` positions due to simplified SFCC.
-        match (
-            self.csu_monitor.get_previous_detent(),
-            self.csu_monitor.get_current_detent(),
-        ) {
-            (CSU::Conf0 | CSU::Conf1, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    < Self::HANDLE_ONE_CONF_AIRSPEED_THRESHOLD_KNOTS
-                    || context.is_on_ground() =>
-            {
-                FlapsConf::Conf1F
-            }
-            (CSU::Conf0 | CSU::Conf1, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::CRUISE_BAULK_AIRSPEED_THRESHOLD_KNOTS
-                    // FIXME use ADRs
-                    || context.pressure_altitude().get::<foot>()
-                        > Self::CRUISE_BAULK_ALTITUDE_THRESHOLD_FEET =>
-            {
-                FlapsConf::Conf0
-            }
-            (CSU::Conf0, CSU::Conf1) => FlapsConf::Conf1,
-            (CSU::Conf1, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::CONF1F_TO_CONF1_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf1
-            }
-            (CSU::Conf1, CSU::Conf1) => self.flaps_conf,
-            (_, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    <= Self::CONF1F_TO_CONF1_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf1F
-            }
-            (_, CSU::Conf1) => FlapsConf::Conf1,
-            (_, CSU::Conf0) if context.is_in_flight() && self.alpha_speed_lock_active => {
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::ALPHA_SPEED_LOCK_OUT_AIRSPEED_THRESHOLD_KNOTS
-                    || self
-                        .angle_of_attack(adirs)
-                        .unwrap_or_default()
-                        .get::<degree>()
-                        < Self::ALPHA_SPEED_LOCK_OUT_AOA_THRESHOLD_DEGREES
-                {
-                    FlapsConf::Conf0
-                } else {
-                    self.flaps_conf
-                }
-            }
-            (CSU::Conf1, CSU::Conf0)
-            | (CSU::Conf2, CSU::Conf0)
-            | (CSU::Conf3, CSU::Conf0)
-            | (CSU::ConfFull, CSU::Conf0)
-                if context.is_in_flight()
-                    && (context.indicated_airspeed().get::<knot>()
-                        < Self::ALPHA_SPEED_LOCK_IN_AIRSPEED_THRESHOLD_KNOTS
-                        || self
-                            .angle_of_attack(adirs)
-                            .unwrap_or_default()
-                            .get::<degree>()
-                            > Self::ALPHA_SPEED_LOCK_IN_AOA_THRESHOLD_DEGREES) =>
-            {
-                FlapsConf::Conf1F
-            }
-            (_, CSU::Conf0) => FlapsConf::Conf0,
-            (CSU::Conf1 | CSU::Conf2, CSU::Conf2)
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::FLRS_CONF2_TO_CONF1F_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf1F
-            }
-            (CSU::Conf2, CSU::Conf2) if self.flaps_conf == FlapsConf::Conf1F => {
-                if context.indicated_airspeed().get::<knot>()
-                    < Self::FLRS_CONF1F_TO_CONF2_AIRSPEED_THRESHOLD_KNOTS
-                {
-                    FlapsConf::Conf2
-                } else {
-                    FlapsConf::Conf1F
-                }
-            }
-            (CSU::Conf2 | CSU::Conf3, CSU::Conf3)
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::FLRS_CONF3_TO_CONF2S_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf2S
-            }
-            (CSU::Conf3, CSU::Conf3) if self.flaps_conf == FlapsConf::Conf2S => {
-                if context.indicated_airspeed().get::<knot>()
-                    < Self::FLRS_CONF2S_TO_CONF3_AIRSPEED_THRESHOLD_KNOTS
-                {
-                    FlapsConf::Conf3
-                } else {
-                    FlapsConf::Conf2S
-                }
-            }
-            (CSU::Conf3 | CSU::ConfFull, CSU::ConfFull)
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::FLRS_CONFFULL_TO_CONF3_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf3
-            }
-            (CSU::ConfFull, CSU::ConfFull) if self.flaps_conf == FlapsConf::Conf3 => {
-                if context.indicated_airspeed().get::<knot>()
-                    < Self::FLRS_CONF3_TO_CONFFULL_AIRSPEED_THRESHOLD_KNOTS
-                {
-                    FlapsConf::ConfFull
-                } else {
-                    FlapsConf::Conf3
-                }
-            }
-            (from, CSU::Conf2) if from != CSU::Conf2 => FlapsConf::Conf2,
-            (from, CSU::Conf3) if from != CSU::Conf3 => FlapsConf::Conf3,
-            (from, CSU::ConfFull) if from != CSU::ConfFull => FlapsConf::ConfFull,
-            (_, _) => self.flaps_conf,
-        }
-    }
-
-    fn flap_load_relief_active(&self) -> bool {
-        let handle_position = self.csu_monitor.get_current_detent();
-        handle_position == CSU::Conf2 && self.flaps_conf != FlapsConf::Conf2
-            || handle_position == CSU::Conf3 && self.flaps_conf != FlapsConf::Conf3
-            || handle_position == CSU::ConfFull && self.flaps_conf != FlapsConf::ConfFull
-    }
-
-    fn cruise_baulk_active(&self) -> bool {
-        let handle_position = self.csu_monitor.get_current_detent();
-        handle_position == CSU::Conf1 && self.flaps_conf == FlapsConf::Conf0
-    }
-
-    fn alpha_speed_lock_active(&self) -> bool {
-        let handle_position = self.csu_monitor.get_current_detent();
-        handle_position == CSU::Conf0
-            && (self.flaps_conf == FlapsConf::Conf1 || self.flaps_conf == FlapsConf::Conf1F)
     }
 
     fn surface_movement_required(demanded_angle: Angle, feedback_angle: Angle) -> bool {
         (demanded_angle - feedback_angle).get::<degree>().abs() > Self::EQUAL_ANGLE_DELTA_DEGREE
     }
 
-    // FIXME This is not the correct ADR input selection yet, due to missing references
-    fn angle_of_attack(&self, adirs: &impl AdirsMeasurementOutputs) -> Option<Angle> {
-        [1, 2, 3]
-            .iter()
-            .find_map(|&adiru_number| adirs.angle_of_attack(adiru_number).normal_value())
+    pub fn calculate_config(&self) -> FlapsConf {
+        let s = self.slats_channel.get_demanded_angle();
+        let f = self.flaps_channel.get_demanded_angle();
+
+        if s == Angle::default() && f == Angle::default() {
+            return FlapsConf::Conf0;
+        } else if s == Angle::new::<degree>(222.27) && f == Angle::default() {
+            return FlapsConf::Conf1;
+        } else if s == Angle::new::<degree>(222.27) && f == Angle::new::<degree>(120.22) {
+            return FlapsConf::Conf1F;
+        } else if s == Angle::new::<degree>(272.27) && f == Angle::new::<degree>(145.51) {
+            return FlapsConf::Conf2;
+        } else if s == Angle::new::<degree>(272.27) && f == Angle::new::<degree>(168.35) {
+            return FlapsConf::Conf3;
+        } else if s == Angle::new::<degree>(334.16) && f == Angle::new::<degree>(251.97) {
+            return FlapsConf::ConfFull;
+        } else {
+            panic!(
+                "Invalid combination sfcc {} {}",
+                s.get::<degree>(),
+                f.get::<degree>()
+            );
+        }
     }
 
     pub fn update(
         &mut self,
         context: &UpdateContext,
-        adirs: &impl AdirsMeasurementOutputs,
         flaps_feedback: &impl PositionPickoffUnit,
         slats_feedback: &impl PositionPickoffUnit,
     ) {
-        self.csu_monitor.update(context);
+        self.flaps_channel.update(context, flaps_feedback);
+        self.slats_channel.update(context, slats_feedback);
 
-        self.flaps_handle_position = self.csu_monitor.get_current_detent();
-        self.flaps_conf = self.generate_configuration(context, adirs);
-        self.flap_load_relief_active = self.flap_load_relief_active();
-        self.cruise_baulk_active = self.cruise_baulk_active();
-        self.alpha_speed_lock_active = self.alpha_speed_lock_active();
-
-        self.flaps_demanded_angle = Self::demanded_flaps_fppu_angle_from_conf(self.flaps_conf);
-        self.slats_demanded_angle = Self::demanded_slats_fppu_angle_from_conf(self.flaps_conf);
-        self.flaps_feedback_angle = flaps_feedback.angle();
-        self.slats_feedback_angle = slats_feedback.angle();
+        self.config = self.calculate_config();
     }
 
     fn slat_flap_system_status_word(&self) -> Arinc429Word<u32> {
@@ -296,16 +101,19 @@ impl SlatFlapControlComputer {
         word.set_bit(14, false);
         word.set_bit(15, false);
         word.set_bit(16, false);
-        word.set_bit(17, self.flaps_handle_position == CSU::Conf0);
-        word.set_bit(18, self.flaps_handle_position == CSU::Conf1);
-        word.set_bit(19, self.flaps_handle_position == CSU::Conf2);
-        word.set_bit(20, self.flaps_handle_position == CSU::Conf3);
-        word.set_bit(21, self.flaps_handle_position == CSU::ConfFull);
-        word.set_bit(22, self.flap_load_relief_active);
+        word.set_bit(17, self.config == FlapsConf::Conf0);
+        word.set_bit(
+            18,
+            self.config == FlapsConf::Conf1 || self.config == FlapsConf::Conf1F,
+        );
+        word.set_bit(19, self.config == FlapsConf::Conf2);
+        word.set_bit(20, self.config == FlapsConf::Conf3);
+        word.set_bit(21, self.config == FlapsConf::ConfFull);
+        word.set_bit(22, false);
         word.set_bit(23, false);
-        word.set_bit(24, self.alpha_speed_lock_active);
-        word.set_bit(25, self.cruise_baulk_active);
-        word.set_bit(26, self.flaps_conf == FlapsConf::Conf1);
+        word.set_bit(24, false);
+        word.set_bit(25, false);
+        word.set_bit(26, self.config == FlapsConf::Conf1);
         word.set_bit(27, false);
         word.set_bit(28, true);
         word.set_bit(29, true);
@@ -317,62 +125,53 @@ impl SlatFlapControlComputer {
         let mut word = Arinc429Word::new(0, SignStatus::NormalOperation);
 
         word.set_bit(11, true);
-        // Slats retracted
         word.set_bit(
             12,
-            self.slats_feedback_angle > Angle::new::<degree>(-5.0)
-                && self.slats_feedback_angle < Angle::new::<degree>(6.2),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(-5.0)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(6.2),
         );
-        // Slats >= 19°
         word.set_bit(
             13,
-            self.slats_feedback_angle > Angle::new::<degree>(234.7)
-                && self.slats_feedback_angle < Angle::new::<degree>(337.),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(210.4)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(337.),
         );
-        // Slats >= 22°
         word.set_bit(
             14,
-            self.slats_feedback_angle > Angle::new::<degree>(272.2)
-                && self.slats_feedback_angle < Angle::new::<degree>(337.),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(321.8)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(337.),
         );
-        // Slats extended 23°
         word.set_bit(
             15,
-            self.slats_feedback_angle > Angle::new::<degree>(280.)
-                && self.slats_feedback_angle < Angle::new::<degree>(337.),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(327.4)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(337.),
         );
         word.set_bit(16, false);
         word.set_bit(17, false);
         word.set_bit(18, true);
-        // Flaps retracted
         word.set_bit(
             19,
-            self.flaps_feedback_angle > Angle::new::<degree>(-5.0)
-                && self.flaps_feedback_angle < Angle::new::<degree>(2.5),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(-5.0)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(2.5),
         );
-        // Flaps >= 7°
         word.set_bit(
             20,
-            self.flaps_feedback_angle > Angle::new::<degree>(102.1)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(140.7)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
-        // Flaps >= 16°
         word.set_bit(
             21,
-            self.flaps_feedback_angle > Angle::new::<degree>(150.0)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(163.7)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
-        // Flaps >= 25°
         word.set_bit(
             22,
-            self.flaps_feedback_angle > Angle::new::<degree>(189.8)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(247.8)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
-        // Flaps extended 32°
         word.set_bit(
             23,
-            self.flaps_feedback_angle > Angle::new::<degree>(218.)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(250.)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
         word.set_bit(24, false);
         word.set_bit(25, false);
@@ -382,20 +181,6 @@ impl SlatFlapControlComputer {
         word.set_bit(29, false);
 
         word
-    }
-
-    fn slat_actual_position_word(&self) -> Arinc429Word<f64> {
-        Arinc429Word::new(
-            self.slats_feedback_angle.get::<degree>(),
-            SignStatus::NormalOperation,
-        )
-    }
-
-    fn flap_actual_position_word(&self) -> Arinc429Word<f64> {
-        Arinc429Word::new(
-            self.flaps_feedback_angle.get::<degree>(),
-            SignStatus::NormalOperation,
-        )
     }
 }
 
@@ -408,19 +193,19 @@ impl SlatFlapLane for SlatFlapControlComputer {
         match surface_type {
             "FLAPS"
                 if Self::surface_movement_required(
-                    self.flaps_demanded_angle,
-                    self.flaps_feedback_angle,
+                    self.flaps_channel.get_demanded_angle(),
+                    self.flaps_channel.get_feedback_angle(),
                 ) =>
             {
-                Some(self.flaps_demanded_angle)
+                Some(self.flaps_channel.get_demanded_angle())
             }
             "SLATS"
                 if Self::surface_movement_required(
-                    self.slats_demanded_angle,
-                    self.slats_feedback_angle,
+                    self.slats_channel.get_demanded_angle(),
+                    self.slats_channel.get_feedback_angle(),
                 ) =>
             {
-                Some(self.slats_demanded_angle)
+                Some(self.slats_channel.get_demanded_angle())
             }
             "FLAPS" | "SLATS" => None,
             _ => panic!("Not a valid slat/flap surface"),
@@ -430,12 +215,13 @@ impl SlatFlapLane for SlatFlapControlComputer {
 
 impl SimulationElement for SlatFlapControlComputer {
     fn accept<T: SimulationElementVisitor>(&mut self, visitor: &mut T) {
-        self.csu_monitor.accept(visitor);
+        self.flaps_channel.accept(visitor);
+        self.slats_channel.accept(visitor);
         visitor.visit(self);
     }
 
     fn write(&self, writer: &mut SimulatorWriter) {
-        writer.write(&self.flaps_conf_index_id, self.flaps_conf as u8);
+        writer.write(&self.config_index_id, self.config as u8);
 
         writer.write(
             &self.slat_flap_system_status_word_id,
@@ -444,14 +230,6 @@ impl SimulationElement for SlatFlapControlComputer {
         writer.write(
             &self.slat_flap_actual_position_word_id,
             self.slat_flap_actual_position_word(),
-        );
-        writer.write(
-            &self.slat_actual_position_word_id,
-            self.slat_actual_position_word(),
-        );
-        writer.write(
-            &self.flap_actual_position_word_id,
-            self.flap_actual_position_word(),
         );
     }
 }
@@ -473,12 +251,11 @@ impl SlatFlapComplex {
     pub fn update(
         &mut self,
         context: &UpdateContext,
-        adirs: &impl AdirsMeasurementOutputs,
         flaps_feedback: &impl PositionPickoffUnit,
         slats_feedback: &impl PositionPickoffUnit,
     ) {
-        self.sfcc[0].update(context, adirs, flaps_feedback, slats_feedback);
-        self.sfcc[1].update(context, adirs, flaps_feedback, slats_feedback);
+        self.sfcc[0].update(context, flaps_feedback, slats_feedback);
+        self.sfcc[1].update(context, flaps_feedback, slats_feedback);
     }
 
     // `idx` 0 is for SFCC1
@@ -605,7 +382,7 @@ mod tests {
         }
     }
 
-    struct A380FlapsTestAircraft {
+    struct A320FlapsTestAircraft {
         green_hydraulic_pressure_id: VariableIdentifier,
         blue_hydraulic_pressure_id: VariableIdentifier,
         yellow_hydraulic_pressure_id: VariableIdentifier,
@@ -617,11 +394,9 @@ mod tests {
         green_pressure: Pressure,
         blue_pressure: Pressure,
         yellow_pressure: Pressure,
-
-        adirs: TestAdirs,
     }
 
-    impl A380FlapsTestAircraft {
+    impl A320FlapsTestAircraft {
         fn new(context: &mut InitContext) -> Self {
             Self {
                 green_hydraulic_pressure_id: context
@@ -648,20 +423,14 @@ mod tests {
                 green_pressure: Pressure::new::<psi>(0.),
                 blue_pressure: Pressure::new::<psi>(0.),
                 yellow_pressure: Pressure::new::<psi>(0.),
-
-                adirs: TestAdirs::new(),
             }
-        }
-
-        fn set_angle_of_attack(&mut self, v: Angle) {
-            self.adirs.set_angle_of_attack(v);
         }
     }
 
-    impl Aircraft for A380FlapsTestAircraft {
+    impl Aircraft for A320FlapsTestAircraft {
         fn update_after_power_distribution(&mut self, context: &UpdateContext) {
             self.slat_flap_complex
-                .update(context, &self.adirs, &self.flap_gear, &self.slat_gear);
+                .update(context, &self.flap_gear, &self.slat_gear);
             self.flap_gear.update(
                 context,
                 &self.slat_flap_complex.sfcc[0],
@@ -677,7 +446,7 @@ mod tests {
         }
     }
 
-    impl SimulationElement for A380FlapsTestAircraft {
+    impl SimulationElement for A320FlapsTestAircraft {
         fn accept<T: SimulationElementVisitor>(&mut self, visitor: &mut T) {
             self.slat_flap_complex.accept(visitor);
             self.flap_gear.accept(visitor);
@@ -692,75 +461,16 @@ mod tests {
         }
     }
 
-    struct TestAdirs {
-        is_aligned: bool,
-        latitude: Arinc429Word<Angle>,
-        longitude: Arinc429Word<Angle>,
-        heading: Arinc429Word<Angle>,
-        vertical_speed: Arinc429Word<Velocity>,
-        altitude: Arinc429Word<Length>,
-        angle_of_attack: Arinc429Word<Angle>,
-    }
-    impl TestAdirs {
-        fn new() -> Self {
-            Self {
-                is_aligned: false,
-                latitude: Arinc429Word::new(Angle::default(), SignStatus::FailureWarning),
-                longitude: Arinc429Word::new(Angle::default(), SignStatus::FailureWarning),
-                heading: Arinc429Word::new(Angle::default(), SignStatus::FailureWarning),
-                vertical_speed: Arinc429Word::new(Velocity::default(), SignStatus::FailureWarning),
-                altitude: Arinc429Word::new(Length::default(), SignStatus::FailureWarning),
-                angle_of_attack: Arinc429Word::new(Angle::default(), SignStatus::NormalOperation),
-            }
-        }
-
-        fn set_angle_of_attack(&mut self, v: Angle) {
-            self.angle_of_attack = Arinc429Word::new(v, SignStatus::NormalOperation);
-        }
-    }
-    impl AdirsMeasurementOutputs for TestAdirs {
-        fn is_fully_aligned(&self, _adiru_number: usize) -> bool {
-            self.is_aligned
-        }
-
-        fn latitude(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
-            self.latitude
-        }
-
-        fn longitude(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
-            self.longitude
-        }
-
-        fn heading(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
-            self.heading
-        }
-
-        fn true_heading(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
-            self.heading
-        }
-
-        fn vertical_speed(&self, _adiru_number: usize) -> Arinc429Word<Velocity> {
-            self.vertical_speed
-        }
-
-        fn altitude(&self, _adiru_number: usize) -> Arinc429Word<Length> {
-            self.altitude
-        }
-
-        fn angle_of_attack(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
-            self.angle_of_attack
-        }
-    }
-    struct A380FlapsTestBed {
-        test_bed: SimulationTestBed<A380FlapsTestAircraft>,
+    struct A320FlapsTestBed {
+        test_bed: SimulationTestBed<A320FlapsTestAircraft>,
     }
 
-    impl A380FlapsTestBed {
+    impl A320FlapsTestBed {
         const HYD_TIME_STEP_MILLIS: u64 = 33;
 
         fn new() -> Self {
             Self {
-                test_bed: SimulationTestBed::new(A380FlapsTestAircraft::new),
+                test_bed: SimulationTestBed::new(A320FlapsTestAircraft::new),
             }
         }
 
@@ -792,18 +502,36 @@ mod tests {
             self.read_by_name(&format!("SFCC_{num}_SLAT_FLAP_ACTUAL_POSITION_WORD"))
         }
 
+        fn read_sfcc_fap_1_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_1"))
+        }
+
+        fn read_sfcc_fap_2_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_2"))
+        }
+
+        fn read_sfcc_fap_3_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_3"))
+        }
+
+        fn read_sfcc_fap_4_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_4"))
+        }
+
+        fn read_sfcc_fap_5_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_5"))
+        }
+
+        fn read_sfcc_fap_6_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_6"))
+        }
+
+        fn read_sfcc_fap_7_word(&mut self, num: u8) -> bool {
+            self.read_by_name(&format!("SFCC_{num}_FAP_7"))
+        }
+
         fn set_indicated_airspeed(mut self, indicated_airspeed: f64) -> Self {
             self.write_by_name("AIRSPEED INDICATED", indicated_airspeed);
-            self
-        }
-
-        fn set_angle_of_attack(mut self, angle_of_attack: Angle) -> Self {
-            self.command(|a| a.set_angle_of_attack(angle_of_attack));
-            self
-        }
-
-        fn set_on_ground(mut self, on_ground: bool) -> Self {
-            self.write_by_name("SIM ON GROUND", on_ground);
             self
         }
 
@@ -825,7 +553,8 @@ mod tests {
         fn get_flaps_demanded_angle(&self, idx: usize) -> f64 {
             self.query(|a| {
                 a.slat_flap_complex.sfcc[idx]
-                    .flaps_demanded_angle
+                    .flaps_channel
+                    .get_demanded_angle()
                     .get::<degree>()
             })
         }
@@ -833,13 +562,14 @@ mod tests {
         fn get_slats_demanded_angle(&self, idx: usize) -> f64 {
             self.query(|a| {
                 a.slat_flap_complex.sfcc[idx]
-                    .slats_demanded_angle
+                    .slats_channel
+                    .get_demanded_angle()
                     .get::<degree>()
             })
         }
 
         fn get_flaps_conf(&self, num: usize) -> FlapsConf {
-            self.query(|a| a.slat_flap_complex.sfcc[num].flaps_conf)
+            self.query(|a| a.slat_flap_complex.sfcc[num].config)
         }
 
         fn get_flaps_fppu_feedback(&self) -> f64 {
@@ -848,6 +578,34 @@ mod tests {
 
         fn get_slats_fppu_feedback(&self) -> f64 {
             self.query(|a| a.slat_gear.angle().get::<degree>())
+        }
+
+        fn get_fap_1(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(0))
+        }
+
+        fn get_fap_2(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(1))
+        }
+
+        fn get_fap_3(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(2))
+        }
+
+        fn get_fap_4(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(3))
+        }
+
+        fn get_fap_5(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(4))
+        }
+
+        fn get_fap_6(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(5))
+        }
+
+        fn get_fap_7(&self, idx: usize) -> bool {
+            self.query(|a| a.slat_flap_complex.sfcc[idx].flaps_channel.get_fap(6))
         }
 
         fn test_flap_conf(
@@ -860,30 +618,31 @@ mod tests {
         ) {
             assert_eq!(self.read_flaps_handle_position(), handle_pos);
             assert!((self.get_flaps_demanded_angle(0) - flaps_demanded_angle).abs() < angle_delta);
-            assert!((self.get_flaps_demanded_angle(1) - flaps_demanded_angle).abs() < angle_delta);
             assert!((self.get_slats_demanded_angle(0) - slats_demanded_angle).abs() < angle_delta);
-            assert!((self.get_slats_demanded_angle(1) - slats_demanded_angle).abs() < angle_delta);
             assert_eq!(self.get_flaps_conf(0), conf);
+
+            assert!((self.get_flaps_demanded_angle(1) - flaps_demanded_angle).abs() < angle_delta);
+            assert!((self.get_slats_demanded_angle(1) - slats_demanded_angle).abs() < angle_delta);
             assert_eq!(self.get_flaps_conf(1), conf);
         }
     }
-    impl TestBed for A380FlapsTestBed {
-        type Aircraft = A380FlapsTestAircraft;
+    impl TestBed for A320FlapsTestBed {
+        type Aircraft = A320FlapsTestAircraft;
 
-        fn test_bed(&self) -> &SimulationTestBed<A380FlapsTestAircraft> {
+        fn test_bed(&self) -> &SimulationTestBed<A320FlapsTestAircraft> {
             &self.test_bed
         }
 
-        fn test_bed_mut(&mut self) -> &mut SimulationTestBed<A380FlapsTestAircraft> {
+        fn test_bed_mut(&mut self) -> &mut SimulationTestBed<A320FlapsTestAircraft> {
             &mut self.test_bed
         }
     }
 
-    fn test_bed() -> A380FlapsTestBed {
-        A380FlapsTestBed::new()
+    fn test_bed() -> A320FlapsTestBed {
+        A320FlapsTestBed::new()
     }
 
-    fn test_bed_with() -> A380FlapsTestBed {
+    fn test_bed_with() -> A320FlapsTestBed {
         test_bed()
     }
 
@@ -903,6 +662,22 @@ mod tests {
 
         assert!(test_bed.contains_variable_with_name("FLAPS_CONF_INDEX"));
 
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_1"));
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_2"));
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_3"));
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_4"));
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_5"));
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_6"));
+        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_7"));
+
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_1"));
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_2"));
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_3"));
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_4"));
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_5"));
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_6"));
+        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_7"));
+
         assert!(test_bed.contains_variable_with_name("SFCC_1_FLAP_ACTUAL_POSITION_WORD"));
         assert!(test_bed.contains_variable_with_name("SFCC_2_FLAP_ACTUAL_POSITION_WORD"));
 
@@ -917,12 +692,199 @@ mod tests {
     }
 
     #[test]
+    fn sfcc_faps() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_yellow_hyd_pressure()
+            .set_blue_hyd_pressure()
+            .set_indicated_airspeed(0.)
+            .set_flaps_handle_position(0)
+            .run_one_tick();
+
+        assert!(!test_bed.get_fap_1(0));
+        assert!(!test_bed.get_fap_2(0));
+        assert!(!test_bed.get_fap_3(0));
+        assert!(test_bed.get_fap_4(0));
+        assert!(!test_bed.get_fap_5(0));
+        assert!(!test_bed.get_fap_6(0));
+        assert!(!test_bed.get_fap_7(0));
+
+        assert!(!test_bed.get_fap_1(1));
+        assert!(!test_bed.get_fap_2(1));
+        assert!(!test_bed.get_fap_3(1));
+        assert!(test_bed.get_fap_4(1));
+        assert!(!test_bed.get_fap_5(1));
+        assert!(!test_bed.get_fap_6(1));
+        assert!(!test_bed.get_fap_7(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(1));
+        assert!(!test_bed.read_sfcc_fap_2_word(1));
+        assert!(!test_bed.read_sfcc_fap_3_word(1));
+        assert!(test_bed.read_sfcc_fap_4_word(1));
+        assert!(!test_bed.read_sfcc_fap_5_word(1));
+        assert!(!test_bed.read_sfcc_fap_6_word(1));
+        assert!(!test_bed.read_sfcc_fap_7_word(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(2));
+        assert!(!test_bed.read_sfcc_fap_2_word(2));
+        assert!(!test_bed.read_sfcc_fap_3_word(2));
+        assert!(test_bed.read_sfcc_fap_4_word(2));
+        assert!(!test_bed.read_sfcc_fap_5_word(2));
+        assert!(!test_bed.read_sfcc_fap_6_word(2));
+        assert!(!test_bed.read_sfcc_fap_7_word(2));
+
+        test_bed = test_bed
+            .set_flaps_handle_position(1)
+            .run_waiting_for(Duration::from_secs(60));
+
+        assert!(!test_bed.get_fap_1(0));
+        assert!(test_bed.get_fap_2(0));
+        assert!(!test_bed.get_fap_3(0));
+        assert!(!test_bed.get_fap_4(0));
+        assert!(!test_bed.get_fap_5(0));
+        assert!(!test_bed.get_fap_6(0));
+        assert!(test_bed.get_fap_7(0));
+
+        assert!(!test_bed.get_fap_1(1));
+        assert!(test_bed.get_fap_2(1));
+        assert!(!test_bed.get_fap_3(1));
+        assert!(!test_bed.get_fap_4(1));
+        assert!(!test_bed.get_fap_5(1));
+        assert!(!test_bed.get_fap_6(1));
+        assert!(test_bed.get_fap_7(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(1));
+        assert!(test_bed.read_sfcc_fap_2_word(1));
+        assert!(!test_bed.read_sfcc_fap_3_word(1));
+        assert!(!test_bed.read_sfcc_fap_4_word(1));
+        assert!(!test_bed.read_sfcc_fap_5_word(1));
+        assert!(!test_bed.read_sfcc_fap_6_word(1));
+        assert!(test_bed.read_sfcc_fap_7_word(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(2));
+        assert!(test_bed.read_sfcc_fap_2_word(2));
+        assert!(!test_bed.read_sfcc_fap_3_word(2));
+        assert!(!test_bed.read_sfcc_fap_4_word(2));
+        assert!(!test_bed.read_sfcc_fap_5_word(2));
+        assert!(!test_bed.read_sfcc_fap_6_word(2));
+        assert!(test_bed.read_sfcc_fap_7_word(2));
+
+        test_bed = test_bed
+            .set_flaps_handle_position(2)
+            .run_waiting_for(Duration::from_secs(60));
+
+        assert!(!test_bed.get_fap_1(0));
+        assert!(test_bed.get_fap_2(0));
+        assert!(!test_bed.get_fap_3(0));
+        assert!(!test_bed.get_fap_4(0));
+        assert!(!test_bed.get_fap_5(0));
+        assert!(!test_bed.get_fap_6(0));
+        assert!(test_bed.get_fap_7(0));
+
+        assert!(!test_bed.get_fap_1(1));
+        assert!(test_bed.get_fap_2(1));
+        assert!(!test_bed.get_fap_3(1));
+        assert!(!test_bed.get_fap_4(1));
+        assert!(!test_bed.get_fap_5(1));
+        assert!(!test_bed.get_fap_6(1));
+        assert!(test_bed.get_fap_7(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(1));
+        assert!(test_bed.read_sfcc_fap_2_word(1));
+        assert!(!test_bed.read_sfcc_fap_3_word(1));
+        assert!(!test_bed.read_sfcc_fap_4_word(1));
+        assert!(!test_bed.read_sfcc_fap_5_word(1));
+        assert!(!test_bed.read_sfcc_fap_6_word(1));
+        assert!(test_bed.read_sfcc_fap_7_word(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(2));
+        assert!(test_bed.read_sfcc_fap_2_word(2));
+        assert!(!test_bed.read_sfcc_fap_3_word(2));
+        assert!(!test_bed.read_sfcc_fap_4_word(2));
+        assert!(!test_bed.read_sfcc_fap_5_word(2));
+        assert!(!test_bed.read_sfcc_fap_6_word(2));
+        assert!(test_bed.read_sfcc_fap_7_word(2));
+
+        test_bed = test_bed
+            .set_flaps_handle_position(3)
+            .run_waiting_for(Duration::from_secs(60));
+
+        assert!(!test_bed.get_fap_1(0));
+        assert!(test_bed.get_fap_2(0));
+        assert!(test_bed.get_fap_3(0));
+        assert!(!test_bed.get_fap_4(0));
+        assert!(test_bed.get_fap_5(0));
+        assert!(!test_bed.get_fap_6(0));
+        assert!(test_bed.get_fap_7(0));
+
+        assert!(!test_bed.get_fap_1(1));
+        assert!(test_bed.get_fap_2(1));
+        assert!(test_bed.get_fap_3(1));
+        assert!(!test_bed.get_fap_4(1));
+        assert!(test_bed.get_fap_5(1));
+        assert!(!test_bed.get_fap_6(1));
+        assert!(test_bed.get_fap_7(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(1));
+        assert!(test_bed.read_sfcc_fap_2_word(1));
+        assert!(test_bed.read_sfcc_fap_3_word(1));
+        assert!(!test_bed.read_sfcc_fap_4_word(1));
+        assert!(test_bed.read_sfcc_fap_5_word(1));
+        assert!(!test_bed.read_sfcc_fap_6_word(1));
+        assert!(test_bed.read_sfcc_fap_7_word(1));
+
+        assert!(!test_bed.read_sfcc_fap_1_word(2));
+        assert!(test_bed.read_sfcc_fap_2_word(2));
+        assert!(test_bed.read_sfcc_fap_3_word(2));
+        assert!(!test_bed.read_sfcc_fap_4_word(2));
+        assert!(test_bed.read_sfcc_fap_5_word(2));
+        assert!(!test_bed.read_sfcc_fap_6_word(2));
+        assert!(test_bed.read_sfcc_fap_7_word(2));
+
+        test_bed = test_bed
+            .set_flaps_handle_position(4)
+            .run_waiting_for(Duration::from_secs(60));
+
+        assert!(test_bed.get_fap_1(0));
+        assert!(test_bed.get_fap_2(0));
+        assert!(test_bed.get_fap_3(0));
+        assert!(!test_bed.get_fap_4(0));
+        assert!(test_bed.get_fap_5(0));
+        assert!(test_bed.get_fap_6(0));
+        assert!(test_bed.get_fap_7(0));
+
+        assert!(test_bed.get_fap_1(1));
+        assert!(test_bed.get_fap_2(1));
+        assert!(test_bed.get_fap_3(1));
+        assert!(!test_bed.get_fap_4(1));
+        assert!(test_bed.get_fap_5(1));
+        assert!(test_bed.get_fap_6(1));
+        assert!(test_bed.get_fap_7(1));
+
+        assert!(test_bed.read_sfcc_fap_1_word(1));
+        assert!(test_bed.read_sfcc_fap_2_word(1));
+        assert!(test_bed.read_sfcc_fap_3_word(1));
+        assert!(!test_bed.read_sfcc_fap_4_word(1));
+        assert!(test_bed.read_sfcc_fap_5_word(1));
+        assert!(test_bed.read_sfcc_fap_6_word(1));
+        assert!(test_bed.read_sfcc_fap_7_word(1));
+
+        assert!(test_bed.read_sfcc_fap_1_word(2));
+        assert!(test_bed.read_sfcc_fap_2_word(2));
+        assert!(test_bed.read_sfcc_fap_3_word(2));
+        assert!(!test_bed.read_sfcc_fap_4_word(2));
+        assert!(test_bed.read_sfcc_fap_5_word(2));
+        assert!(test_bed.read_sfcc_fap_6_word(2));
+        assert!(test_bed.read_sfcc_fap_7_word(2));
+    }
+
+    #[test]
     fn flaps_test_correct_bus_output_clean_config() {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_yellow_hyd_pressure()
             .set_blue_hyd_pressure()
-            .set_indicated_airspeed(160.)
+            .set_indicated_airspeed(0.)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -969,7 +931,7 @@ mod tests {
             .set_green_hyd_pressure()
             .set_yellow_hyd_pressure()
             .set_blue_hyd_pressure()
-            .set_indicated_airspeed(215.)
+            .set_indicated_airspeed(200.)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -987,7 +949,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
@@ -1034,14 +996,14 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(14));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(19));
-        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
+        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(23));
@@ -1051,7 +1013,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(14));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(19));
-        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
+        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(23));
@@ -1081,7 +1043,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
@@ -1089,7 +1051,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
-        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
+        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(23));
 
@@ -1099,7 +1061,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
-        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
+        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(23));
     }
@@ -1128,26 +1090,26 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
-        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(14));
-        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
+        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(14));
+        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
-        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
+        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(23));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(13));
-        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(14));
-        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
+        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(14));
+        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
-        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
+        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(23));
     }
 
@@ -1175,7 +1137,7 @@ mod tests {
         assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(45));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
@@ -1200,65 +1162,65 @@ mod tests {
 
     // Tests flaps configuration and angles for regular
     // increasing handle transitions, i.e 0->1->2->3->4 in sequence
-    // below 205 knots
+    // below 100 knots
     #[test]
-    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_below_205() {
+    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_below_100() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(160.)
+            .set_indicated_airspeed(50.)
             .run_one_tick();
 
         test_bed.test_flap_conf(0, 0., 0., FlapsConf::Conf0, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 108.28, 247.27, FlapsConf::Conf1F, angle_delta);
+        test_bed.test_flap_conf(1, 120.22, 222.27, FlapsConf::Conf1F, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 218.91, 284.65, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
     }
 
     // Tests flaps configuration and angles for regular
     // increasing handle transitions, i.e 0->1->2->3->4 in sequence
-    // above 205 knots
+    // above 100 knots
     #[test]
-    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_above_205() {
+    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_above_100() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(215.)
+            .set_indicated_airspeed(150.)
             .run_one_tick();
 
         test_bed.test_flap_conf(0, 0., 0., FlapsConf::Conf0, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 0., 247.27, FlapsConf::Conf1, angle_delta);
+        test_bed.test_flap_conf(1, 0., 222.27, FlapsConf::Conf1, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 154.65, 284.65, FlapsConf::Conf2S, angle_delta);
+        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
     }
 
-    //Tests regular transition 2->1 below and above 212 knots
+    //Tests regular transition 2->1 below and above 210 knots
     #[test]
     fn flaps_test_regular_handle_transition_pos_2_to_1() {
         let mut test_bed = test_bed_with()
@@ -1288,7 +1250,7 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1);
     }
 
-    //Tests transition between Conf1F to Conf1 above 212 knots
+    //Tests transition between Conf1F to Conf1 above 210 knots
     #[test]
     fn flaps_test_regular_handle_transition_pos_1_to_1() {
         let mut test_bed = test_bed_with()
@@ -1313,30 +1275,30 @@ mod tests {
 
     // Tests flaps configuration and angles for regular
     // decreasing handle transitions, i.e 4->3->2->1->0 in sequence
-    // below 212 knots
+    // below 210 knots
     #[test]
-    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_below_212() {
+    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_below_210() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(160.)
+            .set_indicated_airspeed(150.)
             .run_one_tick();
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 218.91, 284.65, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 108.28, 247.27, FlapsConf::Conf1F, angle_delta);
+        test_bed.test_flap_conf(1, 120.22, 222.27, FlapsConf::Conf1F, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
 
@@ -1347,7 +1309,7 @@ mod tests {
     // decreasing handle transitions, i.e 4->3->2->1->0 in sequence
     // above 210 knots
     #[test]
-    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_above_212() {
+    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_above_210() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
@@ -1356,19 +1318,19 @@ mod tests {
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 218.91, 284.65, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 0., 247.27, FlapsConf::Conf1, angle_delta);
+        test_bed.test_flap_conf(1, 0., 222.27, FlapsConf::Conf1, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
 
@@ -1393,23 +1355,23 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf2);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf3);
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf3);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::ConfFull);
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::ConfFull);
 
         test_bed = test_bed
-            .set_indicated_airspeed(160.)
+            .set_indicated_airspeed(110.)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1484,7 +1446,7 @@ mod tests {
 
         test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(210.)
+            .set_indicated_airspeed(110.)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1501,7 +1463,7 @@ mod tests {
 
         test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(205.)
+            .set_indicated_airspeed(110.)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1580,8 +1542,8 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf2);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf2);
@@ -1592,7 +1554,7 @@ mod tests {
     fn flaps_test_irregular_handle_transition_init_pos_3() {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(205.)
+            .set_indicated_airspeed(150.)
             .set_flaps_handle_position(3)
             .run_one_tick();
 
@@ -1634,8 +1596,8 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf3);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf3);
@@ -1646,7 +1608,7 @@ mod tests {
     fn flaps_test_irregular_handle_transition_init_pos_4() {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(205.)
+            .set_indicated_airspeed(150.)
             .set_flaps_handle_position(4)
             .run_one_tick();
 
@@ -1688,8 +1650,8 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::ConfFull);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::ConfFull);
@@ -1710,7 +1672,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1737,7 +1698,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1746,7 +1706,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(2)
-            .run_waiting_for(Duration::from_secs(35));
+            .run_waiting_for(Duration::from_secs(20));
 
         assert!(
             (test_bed.get_flaps_fppu_feedback() - test_bed.get_flaps_demanded_angle(0)).abs()
@@ -1764,7 +1724,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(2)
             .run_one_tick();
 
@@ -1791,7 +1750,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(3)
             .run_waiting_for(Duration::from_secs(20));
 
@@ -1818,7 +1776,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1827,7 +1784,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(1)
-            .run_waiting_for(Duration::from_secs(60));
+            .run_waiting_for(Duration::from_secs(30));
 
         assert!(
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(0)).abs()
@@ -1854,7 +1811,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(1)
-            .run_waiting_for(Duration::from_secs(60));
+            .run_waiting_for(Duration::from_secs(30));
 
         assert!(
             (test_bed.get_flaps_fppu_feedback() - test_bed.get_flaps_demanded_angle(0)).abs()
@@ -1880,7 +1837,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1889,7 +1845,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(2)
-            .run_waiting_for(Duration::from_secs(60));
+            .run_waiting_for(Duration::from_secs(40));
 
         assert!(
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(0)).abs()
@@ -1907,7 +1863,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(2)
             .run_one_tick();
 
@@ -1916,7 +1871,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(3)
-            .run_waiting_for(Duration::from_secs(60));
+            .run_waiting_for(Duration::from_secs(40));
 
         assert!(
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(0)).abs()
@@ -1934,7 +1889,6 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
-            .set_on_ground(true)
             .set_flaps_handle_position(3)
             .run_one_tick();
 
@@ -1953,134 +1907,5 @@ mod tests {
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(1)).abs()
                 <= angle_delta
         );
-    }
-
-    #[test]
-    fn config_test_flrs_full() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_indicated_airspeed(186.)
-            .set_flaps_handle_position(0)
-            .run_one_tick();
-
-        test_bed = test_bed
-            .set_flaps_handle_position(4)
-            .run_waiting_for(Duration::from_secs(50));
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf3);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf3);
-    }
-
-    #[test]
-    fn config_test_flrs_3() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_indicated_airspeed(200.)
-            .set_flaps_handle_position(0)
-            .run_one_tick();
-
-        test_bed = test_bed
-            .set_flaps_handle_position(3)
-            .run_waiting_for(Duration::from_secs(50));
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf2S);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf2S);
-    }
-
-    #[test]
-    fn config_test_flrs_2() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_indicated_airspeed(224.)
-            .set_flaps_handle_position(0)
-            .run_one_tick();
-
-        test_bed = test_bed
-            .set_flaps_handle_position(2)
-            .run_waiting_for(Duration::from_secs(50));
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
-    }
-
-    #[test]
-    fn config_test_cruise_baulk() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_indicated_airspeed(280.)
-            .set_on_ground(false)
-            .set_flaps_handle_position(0)
-            .run_one_tick();
-
-        test_bed = test_bed
-            .set_flaps_handle_position(1)
-            .run_waiting_for(Duration::from_secs(20));
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
-        assert!(test_bed.read_slat_flap_system_status_word(1).get_bit(25));
-        assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(25));
-    }
-
-    #[test]
-    fn config_test_alpha_lock_speed() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_indicated_airspeed(130.)
-            .set_angle_of_attack(Angle::new::<degree>(5.))
-            .set_on_ground(false)
-            .set_flaps_handle_position(1)
-            .run_one_tick();
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
-        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
-        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
-
-        test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
-        assert!(test_bed.read_slat_flap_system_status_word(1).get_bit(24));
-        assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(24));
-
-        test_bed = test_bed.set_indicated_airspeed(200.).run_one_tick();
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
-        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
-        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
-    }
-
-    #[test]
-    fn config_test_alpha_lock_aoa() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_indicated_airspeed(200.)
-            .set_angle_of_attack(Angle::new::<degree>(10.))
-            .set_on_ground(false)
-            .set_flaps_handle_position(1)
-            .run_one_tick();
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
-        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
-        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
-
-        test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
-        assert!(test_bed.read_slat_flap_system_status_word(1).get_bit(24));
-        assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(24));
-
-        test_bed = test_bed
-            .set_angle_of_attack(Angle::new::<degree>(9.1))
-            .run_one_tick();
-
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
-        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
-        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
     }
 }
