+++ fbw-a380x/src/wasm/systems/a380_systems/src/hydraulic/sfcc.rs
@@ -1,7 +1,7 @@
 use crate::systems::shared::arinc429::{Arinc429Word, SignStatus};
 use systems::accept_iterable;
-use systems::hydraulic::command_sensor_unit::{CSUMonitor, CSU};
-use systems::shared::PositionPickoffUnit;
+use systems::hydraulic::command_sensor_unit::CSU;
+use systems::shared::{AdirsMeasurementOutputs, PositionPickoffUnit};
 
 use systems::simulation::{
     InitContext, SimulationElement, SimulationElementVisitor, SimulatorWriter, UpdateContext,
@@ -9,7 +9,10 @@ use systems::simulation::{
 };
 
 use std::panic;
-use uom::si::{angle::degree, f64::*, velocity::knot};
+use uom::si::{angle::degree, f64::*};
+
+use super::flaps_channel::FlapsChannel;
+use super::slats_channel::SlatsChannel;
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 enum FlapsConf {
@@ -17,121 +20,65 @@ enum FlapsConf {
     Conf1,
     Conf1F,
     Conf2,
+    Conf2S,
     Conf3,
     ConfFull,
 }
 
 struct SlatFlapControlComputer {
-    flaps_conf_index_id: VariableIdentifier,
+    config_index_id: VariableIdentifier,
     slat_flap_system_status_word_id: VariableIdentifier,
     slat_flap_actual_position_word_id: VariableIdentifier,
-    slat_actual_position_word_id: VariableIdentifier,
-    flap_actual_position_word_id: VariableIdentifier,
-    fap_ids: [VariableIdentifier; 7],
 
-    flaps_demanded_angle: Angle,
-    slats_demanded_angle: Angle,
-    flaps_feedback_angle: Angle,
-    slats_feedback_angle: Angle,
-    flaps_conf: FlapsConf,
-    fap: [bool; 7],
-    csu_monitor: CSUMonitor,
+    flaps_channel: FlapsChannel,
+    slats_channel: SlatsChannel,
+
+    config: FlapsConf,
 }
 
 impl SlatFlapControlComputer {
     const EQUAL_ANGLE_DELTA_DEGREE: f64 = 0.177;
-    const HANDLE_ONE_CONF_AIRSPEED_THRESHOLD_KNOTS: f64 = 100.;
-    const CONF1F_TO_CONF1_AIRSPEED_THRESHOLD_KNOTS: f64 = 210.;
 
     fn new(context: &mut InitContext, num: u8) -> Self {
         Self {
-            flaps_conf_index_id: context.get_identifier("FLAPS_CONF_INDEX".to_owned()),
+            config_index_id: context.get_identifier("FLAPS_CONF_INDEX".to_owned()),
             slat_flap_system_status_word_id: context
                 .get_identifier(format!("SFCC_{num}_SLAT_FLAP_SYSTEM_STATUS_WORD")),
             slat_flap_actual_position_word_id: context
                 .get_identifier(format!("SFCC_{num}_SLAT_FLAP_ACTUAL_POSITION_WORD")),
-            slat_actual_position_word_id: context
-                .get_identifier(format!("SFCC_{num}_SLAT_ACTUAL_POSITION_WORD")),
-            flap_actual_position_word_id: context
-                .get_identifier(format!("SFCC_{num}_FLAP_ACTUAL_POSITION_WORD")),
-            fap_ids: [
-                context.get_identifier(format!("SFCC_{num}_FAP_1")),
-                context.get_identifier(format!("SFCC_{num}_FAP_2")),
-                context.get_identifier(format!("SFCC_{num}_FAP_3")),
-                context.get_identifier(format!("SFCC_{num}_FAP_4")),
-                context.get_identifier(format!("SFCC_{num}_FAP_5")),
-                context.get_identifier(format!("SFCC_{num}_FAP_6")),
-                context.get_identifier(format!("SFCC_{num}_FAP_7")),
-            ],
 
-            flaps_demanded_angle: Angle::new::<degree>(0.),
-            slats_demanded_angle: Angle::new::<degree>(0.),
-            flaps_feedback_angle: Angle::new::<degree>(0.),
-            slats_feedback_angle: Angle::new::<degree>(0.),
-            flaps_conf: FlapsConf::Conf0,
+            flaps_channel: FlapsChannel::new(context, num),
+            slats_channel: SlatsChannel::new(context, num),
 
-            // Set to false to match power-off state
-            fap: [false; 7],
-            csu_monitor: CSUMonitor::new(context),
+            config: FlapsConf::Conf0,
         }
     }
 
-    // Returns a flap demanded angle in FPPU reference degree (feedback sensor)
-    fn demanded_flaps_fppu_angle_from_conf(flap_conf: FlapsConf) -> Angle {
-        match flap_conf {
-            FlapsConf::Conf0 => Angle::new::<degree>(0.),
-            FlapsConf::Conf1 => Angle::new::<degree>(0.),
-            FlapsConf::Conf1F => Angle::new::<degree>(120.22),
-            FlapsConf::Conf2 => Angle::new::<degree>(145.51),
-            FlapsConf::Conf3 => Angle::new::<degree>(168.35),
-            FlapsConf::ConfFull => Angle::new::<degree>(251.97),
-        }
-    }
+    // Interpolated from A320 FPPU references, assuming we're using the A320 FPPU
+    pub fn calculate_config(&self) -> FlapsConf {
+        let s = self.slats_channel.get_demanded_angle();
+        let f = self.flaps_channel.get_demanded_angle();
 
-    // Returns a slat demanded angle in FPPU reference degree (feedback sensor)
-    fn demanded_slats_fppu_angle_from_conf(flap_conf: FlapsConf) -> Angle {
-        match flap_conf {
-            FlapsConf::Conf0 => Angle::new::<degree>(0.),
-            FlapsConf::Conf1 => Angle::new::<degree>(222.27),
-            FlapsConf::Conf1F => Angle::new::<degree>(222.27),
-            FlapsConf::Conf2 => Angle::new::<degree>(272.27),
-            FlapsConf::Conf3 => Angle::new::<degree>(272.27),
-            FlapsConf::ConfFull => Angle::new::<degree>(334.16),
-        }
-    }
-
-    fn generate_configuration(&self, context: &UpdateContext) -> FlapsConf {
-        // Ignored `CSU::OutOfDetent` and `CSU::Fault` positions due to simplified SFCC.
-        match (
-            self.csu_monitor.get_previous_detent(),
-            self.csu_monitor.get_current_detent(),
-        ) {
-            (CSU::Conf0, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    <= Self::HANDLE_ONE_CONF_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf1F
-            }
-            (CSU::Conf0, CSU::Conf1) => FlapsConf::Conf1,
-            (CSU::Conf1, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    > Self::CONF1F_TO_CONF1_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf1
-            }
-            (CSU::Conf1, CSU::Conf1) => self.flaps_conf,
-            (_, CSU::Conf1)
-                if context.indicated_airspeed().get::<knot>()
-                    <= Self::CONF1F_TO_CONF1_AIRSPEED_THRESHOLD_KNOTS =>
-            {
-                FlapsConf::Conf1F
-            }
-            (_, CSU::Conf1) => FlapsConf::Conf1,
-            (_, CSU::Conf0) => FlapsConf::Conf0,
-            (from, CSU::Conf2) if from != CSU::Conf2 => FlapsConf::Conf2,
-            (from, CSU::Conf3) if from != CSU::Conf3 => FlapsConf::Conf3,
-            (from, CSU::ConfFull) if from != CSU::ConfFull => FlapsConf::ConfFull,
-            (_, _) => self.flaps_conf,
+        if s == Angle::default() && f == Angle::default() {
+            return FlapsConf::Conf0;
+        } else if s == Angle::new::<degree>(247.27) && f == Angle::default() {
+            return FlapsConf::Conf1;
+        } else if s == Angle::new::<degree>(247.27) && f == Angle::new::<degree>(108.28) {
+            return FlapsConf::Conf1F;
+        } else if s == Angle::new::<degree>(247.27) && f == Angle::new::<degree>(154.65) {
+            return FlapsConf::Conf2;
+        } else if s == Angle::new::<degree>(284.65) && f == Angle::new::<degree>(154.65) {
+            return FlapsConf::Conf2S;
+        } else if s == Angle::new::<degree>(284.65) && f == Angle::new::<degree>(194.03) {
+            return FlapsConf::Conf3;
+        } else if s == Angle::new::<degree>(284.65) && f == Angle::new::<degree>(218.91) {
+            return FlapsConf::ConfFull;
+        } else {
+            panic!(
+                "Invalid combination sfcc {} {}",
+                s.get::<degree>(),
+                f.get::<degree>()
+            );
         }
     }
 
@@ -142,31 +89,14 @@ impl SlatFlapControlComputer {
     pub fn update(
         &mut self,
         context: &UpdateContext,
+        adirs: &impl AdirsMeasurementOutputs,
         flaps_feedback: &impl PositionPickoffUnit,
         slats_feedback: &impl PositionPickoffUnit,
     ) {
-        self.csu_monitor.update(context);
+        self.flaps_channel.update(context, flaps_feedback, adirs);
+        self.slats_channel.update(context, slats_feedback, adirs);
 
-        self.flaps_conf = self.generate_configuration(context);
-
-        self.flaps_demanded_angle = Self::demanded_flaps_fppu_angle_from_conf(self.flaps_conf);
-        self.slats_demanded_angle = Self::demanded_slats_fppu_angle_from_conf(self.flaps_conf);
-        self.flaps_feedback_angle = flaps_feedback.angle();
-        self.slats_feedback_angle = slats_feedback.angle();
-
-        self.fap_update();
-    }
-
-    fn fap_update(&mut self) {
-        let fppu_angle = self.flaps_feedback_angle.get::<degree>();
-
-        self.fap[0] = fppu_angle > 247.8 && fppu_angle < 254.0;
-        self.fap[1] = fppu_angle > 114.6 && fppu_angle < 254.0;
-        self.fap[2] = fppu_angle > 163.7 && fppu_angle < 254.0;
-        self.fap[3] = self.csu_monitor.get_current_detent() == CSU::Conf0;
-        self.fap[4] = fppu_angle > 163.7 && fppu_angle < 254.0;
-        self.fap[5] = fppu_angle > 247.8 && fppu_angle < 254.0;
-        self.fap[6] = fppu_angle > 114.6 && fppu_angle < 254.0;
+        self.config = self.calculate_config();
     }
 
     fn slat_flap_system_status_word(&self) -> Arinc429Word<u32> {
@@ -178,19 +108,16 @@ impl SlatFlapControlComputer {
         word.set_bit(14, false);
         word.set_bit(15, false);
         word.set_bit(16, false);
-        word.set_bit(17, self.flaps_conf == FlapsConf::Conf0);
-        word.set_bit(
-            18,
-            self.flaps_conf == FlapsConf::Conf1 || self.flaps_conf == FlapsConf::Conf1F,
-        );
-        word.set_bit(19, self.flaps_conf == FlapsConf::Conf2);
-        word.set_bit(20, self.flaps_conf == FlapsConf::Conf3);
-        word.set_bit(21, self.flaps_conf == FlapsConf::ConfFull);
-        word.set_bit(22, false);
+        word.set_bit(17, self.flaps_channel.get_handle_detent() == CSU::Conf0);
+        word.set_bit(18, self.flaps_channel.get_handle_detent() == CSU::Conf1);
+        word.set_bit(19, self.flaps_channel.get_handle_detent() == CSU::Conf2);
+        word.set_bit(20, self.flaps_channel.get_handle_detent() == CSU::Conf3);
+        word.set_bit(21, self.flaps_channel.get_handle_detent() == CSU::ConfFull);
+        word.set_bit(22, self.flaps_channel.get_flap_load_relief());
         word.set_bit(23, false);
-        word.set_bit(24, false);
-        word.set_bit(25, false);
-        word.set_bit(26, self.flaps_conf == FlapsConf::Conf1);
+        word.set_bit(24, self.slats_channel.get_alpha_speed_lock());
+        word.set_bit(25, self.slats_channel.get_cruise_baulk());
+        word.set_bit(26, self.config == FlapsConf::Conf1);
         word.set_bit(27, false);
         word.set_bit(28, true);
         word.set_bit(29, true);
@@ -202,53 +129,62 @@ impl SlatFlapControlComputer {
         let mut word = Arinc429Word::new(0, SignStatus::NormalOperation);
 
         word.set_bit(11, true);
+        // Slats retracted
         word.set_bit(
             12,
-            self.slats_feedback_angle > Angle::new::<degree>(-5.0)
-                && self.slats_feedback_angle < Angle::new::<degree>(6.2),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(-5.0)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(6.2),
         );
+        // Slats >= 19°
         word.set_bit(
             13,
-            self.slats_feedback_angle > Angle::new::<degree>(210.4)
-                && self.slats_feedback_angle < Angle::new::<degree>(337.),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(234.7)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(337.),
         );
+        // Slats >= 22°
         word.set_bit(
             14,
-            self.slats_feedback_angle > Angle::new::<degree>(321.8)
-                && self.slats_feedback_angle < Angle::new::<degree>(337.),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(272.2)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(337.),
         );
+        // Slats extended 23°
         word.set_bit(
             15,
-            self.slats_feedback_angle > Angle::new::<degree>(327.4)
-                && self.slats_feedback_angle < Angle::new::<degree>(337.),
+            self.slats_channel.get_feedback_angle() > Angle::new::<degree>(280.)
+                && self.slats_channel.get_feedback_angle() < Angle::new::<degree>(337.),
         );
         word.set_bit(16, false);
         word.set_bit(17, false);
         word.set_bit(18, true);
+        // Flaps retracted
         word.set_bit(
             19,
-            self.flaps_feedback_angle > Angle::new::<degree>(-5.0)
-                && self.flaps_feedback_angle < Angle::new::<degree>(2.5),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(-5.0)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(2.5),
         );
+        // Flaps >= 7°
         word.set_bit(
             20,
-            self.flaps_feedback_angle > Angle::new::<degree>(140.7)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(102.1)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
+        // Flaps >= 16°
         word.set_bit(
             21,
-            self.flaps_feedback_angle > Angle::new::<degree>(163.7)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(150.0)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
+        // Flaps >= 25°
         word.set_bit(
             22,
-            self.flaps_feedback_angle > Angle::new::<degree>(247.8)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(189.8)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
+        // Flaps extended 32°
         word.set_bit(
             23,
-            self.flaps_feedback_angle > Angle::new::<degree>(250.)
-                && self.flaps_feedback_angle < Angle::new::<degree>(254.),
+            self.flaps_channel.get_feedback_angle() > Angle::new::<degree>(218.)
+                && self.flaps_channel.get_feedback_angle() < Angle::new::<degree>(254.),
         );
         word.set_bit(24, false);
         word.set_bit(25, false);
@@ -258,20 +194,6 @@ impl SlatFlapControlComputer {
         word.set_bit(29, false);
 
         word
-    }
-
-    fn slat_actual_position_word(&self) -> Arinc429Word<f64> {
-        Arinc429Word::new(
-            self.slats_feedback_angle.get::<degree>(),
-            SignStatus::NormalOperation,
-        )
-    }
-
-    fn flap_actual_position_word(&self) -> Arinc429Word<f64> {
-        Arinc429Word::new(
-            self.flaps_feedback_angle.get::<degree>(),
-            SignStatus::NormalOperation,
-        )
     }
 }
 
@@ -284,19 +206,19 @@ impl SlatFlapLane for SlatFlapControlComputer {
         match surface_type {
             "FLAPS"
                 if Self::surface_movement_required(
-                    self.flaps_demanded_angle,
-                    self.flaps_feedback_angle,
+                    self.flaps_channel.get_demanded_angle(),
+                    self.flaps_channel.get_feedback_angle(),
                 ) =>
             {
-                Some(self.flaps_demanded_angle)
+                Some(self.flaps_channel.get_demanded_angle())
             }
             "SLATS"
                 if Self::surface_movement_required(
-                    self.slats_demanded_angle,
-                    self.slats_feedback_angle,
+                    self.slats_channel.get_demanded_angle(),
+                    self.slats_channel.get_feedback_angle(),
                 ) =>
             {
-                Some(self.slats_demanded_angle)
+                Some(self.slats_channel.get_demanded_angle())
             }
             "FLAPS" | "SLATS" => None,
             _ => panic!("Not a valid slat/flap surface"),
@@ -306,16 +228,13 @@ impl SlatFlapLane for SlatFlapControlComputer {
 
 impl SimulationElement for SlatFlapControlComputer {
     fn accept<T: SimulationElementVisitor>(&mut self, visitor: &mut T) {
-        self.csu_monitor.accept(visitor);
+        self.flaps_channel.accept(visitor);
+        self.slats_channel.accept(visitor);
         visitor.visit(self);
     }
 
     fn write(&self, writer: &mut SimulatorWriter) {
-        for (id, fap) in self.fap_ids.iter().zip(self.fap) {
-            writer.write(id, fap);
-        }
-
-        writer.write(&self.flaps_conf_index_id, self.flaps_conf as u8);
+        writer.write(&self.config_index_id, self.config as u8);
 
         writer.write(
             &self.slat_flap_system_status_word_id,
@@ -324,14 +243,6 @@ impl SimulationElement for SlatFlapControlComputer {
         writer.write(
             &self.slat_flap_actual_position_word_id,
             self.slat_flap_actual_position_word(),
-        );
-        writer.write(
-            &self.slat_actual_position_word_id,
-            self.slat_actual_position_word(),
-        );
-        writer.write(
-            &self.flap_actual_position_word_id,
-            self.flap_actual_position_word(),
         );
     }
 }
@@ -353,11 +264,12 @@ impl SlatFlapComplex {
     pub fn update(
         &mut self,
         context: &UpdateContext,
+        adirs: &impl AdirsMeasurementOutputs,
         flaps_feedback: &impl PositionPickoffUnit,
         slats_feedback: &impl PositionPickoffUnit,
     ) {
-        self.sfcc[0].update(context, flaps_feedback, slats_feedback);
-        self.sfcc[1].update(context, flaps_feedback, slats_feedback);
+        self.sfcc[0].update(context, adirs, flaps_feedback, slats_feedback);
+        self.sfcc[1].update(context, adirs, flaps_feedback, slats_feedback);
     }
 
     // `idx` 0 is for SFCC1
@@ -484,7 +396,7 @@ mod tests {
         }
     }
 
-    struct A320FlapsTestAircraft {
+    struct A380FlapsTestAircraft {
         green_hydraulic_pressure_id: VariableIdentifier,
         blue_hydraulic_pressure_id: VariableIdentifier,
         yellow_hydraulic_pressure_id: VariableIdentifier,
@@ -496,9 +408,11 @@ mod tests {
         green_pressure: Pressure,
         blue_pressure: Pressure,
         yellow_pressure: Pressure,
+
+        adirs: TestAdirs,
     }
 
-    impl A320FlapsTestAircraft {
+    impl A380FlapsTestAircraft {
         fn new(context: &mut InitContext) -> Self {
             Self {
                 green_hydraulic_pressure_id: context
@@ -525,14 +439,20 @@ mod tests {
                 green_pressure: Pressure::new::<psi>(0.),
                 blue_pressure: Pressure::new::<psi>(0.),
                 yellow_pressure: Pressure::new::<psi>(0.),
+
+                adirs: TestAdirs::new(),
             }
+        }
+
+        fn set_angle_of_attack(&mut self, v: Angle) {
+            self.adirs.set_angle_of_attack(v);
         }
     }
 
-    impl Aircraft for A320FlapsTestAircraft {
+    impl Aircraft for A380FlapsTestAircraft {
         fn update_after_power_distribution(&mut self, context: &UpdateContext) {
             self.slat_flap_complex
-                .update(context, &self.flap_gear, &self.slat_gear);
+                .update(context, &self.adirs, &self.flap_gear, &self.slat_gear);
             self.flap_gear.update(
                 context,
                 &self.slat_flap_complex.sfcc[0],
@@ -548,7 +468,7 @@ mod tests {
         }
     }
 
-    impl SimulationElement for A320FlapsTestAircraft {
+    impl SimulationElement for A380FlapsTestAircraft {
         fn accept<T: SimulationElementVisitor>(&mut self, visitor: &mut T) {
             self.slat_flap_complex.accept(visitor);
             self.flap_gear.accept(visitor);
@@ -563,16 +483,75 @@ mod tests {
         }
     }
 
-    struct A320FlapsTestBed {
-        test_bed: SimulationTestBed<A320FlapsTestAircraft>,
+    struct TestAdirs {
+        is_aligned: bool,
+        latitude: Arinc429Word<Angle>,
+        longitude: Arinc429Word<Angle>,
+        heading: Arinc429Word<Angle>,
+        vertical_speed: Arinc429Word<Velocity>,
+        altitude: Arinc429Word<Length>,
+        angle_of_attack: Arinc429Word<Angle>,
+    }
+    impl TestAdirs {
+        fn new() -> Self {
+            Self {
+                is_aligned: false,
+                latitude: Arinc429Word::new(Angle::default(), SignStatus::FailureWarning),
+                longitude: Arinc429Word::new(Angle::default(), SignStatus::FailureWarning),
+                heading: Arinc429Word::new(Angle::default(), SignStatus::FailureWarning),
+                vertical_speed: Arinc429Word::new(Velocity::default(), SignStatus::FailureWarning),
+                altitude: Arinc429Word::new(Length::default(), SignStatus::FailureWarning),
+                angle_of_attack: Arinc429Word::new(Angle::default(), SignStatus::NormalOperation),
+            }
+        }
+
+        fn set_angle_of_attack(&mut self, v: Angle) {
+            self.angle_of_attack = Arinc429Word::new(v, SignStatus::NormalOperation);
+        }
+    }
+    impl AdirsMeasurementOutputs for TestAdirs {
+        fn is_fully_aligned(&self, _adiru_number: usize) -> bool {
+            self.is_aligned
+        }
+
+        fn latitude(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
+            self.latitude
+        }
+
+        fn longitude(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
+            self.longitude
+        }
+
+        fn heading(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
+            self.heading
+        }
+
+        fn true_heading(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
+            self.heading
+        }
+
+        fn vertical_speed(&self, _adiru_number: usize) -> Arinc429Word<Velocity> {
+            self.vertical_speed
+        }
+
+        fn altitude(&self, _adiru_number: usize) -> Arinc429Word<Length> {
+            self.altitude
+        }
+
+        fn angle_of_attack(&self, _adiru_number: usize) -> Arinc429Word<Angle> {
+            self.angle_of_attack
+        }
+    }
+    struct A380FlapsTestBed {
+        test_bed: SimulationTestBed<A380FlapsTestAircraft>,
     }
 
-    impl A320FlapsTestBed {
+    impl A380FlapsTestBed {
         const HYD_TIME_STEP_MILLIS: u64 = 33;
 
         fn new() -> Self {
             Self {
-                test_bed: SimulationTestBed::new(A320FlapsTestAircraft::new),
+                test_bed: SimulationTestBed::new(A380FlapsTestAircraft::new),
             }
         }
 
@@ -604,36 +583,18 @@ mod tests {
             self.read_by_name(&format!("SFCC_{num}_SLAT_FLAP_ACTUAL_POSITION_WORD"))
         }
 
-        fn read_sfcc_fap_1_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_1"))
-        }
-
-        fn read_sfcc_fap_2_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_2"))
-        }
-
-        fn read_sfcc_fap_3_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_3"))
-        }
-
-        fn read_sfcc_fap_4_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_4"))
-        }
-
-        fn read_sfcc_fap_5_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_5"))
-        }
-
-        fn read_sfcc_fap_6_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_6"))
-        }
-
-        fn read_sfcc_fap_7_word(&mut self, num: u8) -> bool {
-            self.read_by_name(&format!("SFCC_{num}_FAP_7"))
-        }
-
         fn set_indicated_airspeed(mut self, indicated_airspeed: f64) -> Self {
             self.write_by_name("AIRSPEED INDICATED", indicated_airspeed);
+            self
+        }
+
+        fn set_angle_of_attack(mut self, angle_of_attack: Angle) -> Self {
+            self.command(|a| a.set_angle_of_attack(angle_of_attack));
+            self
+        }
+
+        fn set_on_ground(mut self, on_ground: bool) -> Self {
+            self.write_by_name("SIM ON GROUND", on_ground);
             self
         }
 
@@ -655,7 +616,8 @@ mod tests {
         fn get_flaps_demanded_angle(&self, idx: usize) -> f64 {
             self.query(|a| {
                 a.slat_flap_complex.sfcc[idx]
-                    .flaps_demanded_angle
+                    .flaps_channel
+                    .get_demanded_angle()
                     .get::<degree>()
             })
         }
@@ -663,13 +625,14 @@ mod tests {
         fn get_slats_demanded_angle(&self, idx: usize) -> f64 {
             self.query(|a| {
                 a.slat_flap_complex.sfcc[idx]
-                    .slats_demanded_angle
+                    .slats_channel
+                    .get_demanded_angle()
                     .get::<degree>()
             })
         }
 
         fn get_flaps_conf(&self, num: usize) -> FlapsConf {
-            self.query(|a| a.slat_flap_complex.sfcc[num].flaps_conf)
+            self.query(|a| a.slat_flap_complex.sfcc[num].config)
         }
 
         fn get_flaps_fppu_feedback(&self) -> f64 {
@@ -678,34 +641,6 @@ mod tests {
 
         fn get_slats_fppu_feedback(&self) -> f64 {
             self.query(|a| a.slat_gear.angle().get::<degree>())
-        }
-
-        fn get_fap_1(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[0])
-        }
-
-        fn get_fap_2(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[1])
-        }
-
-        fn get_fap_3(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[2])
-        }
-
-        fn get_fap_4(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[3])
-        }
-
-        fn get_fap_5(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[4])
-        }
-
-        fn get_fap_6(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[5])
-        }
-
-        fn get_fap_7(&self, idx: usize) -> bool {
-            self.query(|a| a.slat_flap_complex.sfcc[idx].fap[6])
         }
 
         fn test_flap_conf(
@@ -718,31 +653,30 @@ mod tests {
         ) {
             assert_eq!(self.read_flaps_handle_position(), handle_pos);
             assert!((self.get_flaps_demanded_angle(0) - flaps_demanded_angle).abs() < angle_delta);
-            assert!((self.get_slats_demanded_angle(0) - slats_demanded_angle).abs() < angle_delta);
-            assert_eq!(self.get_flaps_conf(0), conf);
-
             assert!((self.get_flaps_demanded_angle(1) - flaps_demanded_angle).abs() < angle_delta);
+            assert!((self.get_slats_demanded_angle(0) - slats_demanded_angle).abs() < angle_delta);
             assert!((self.get_slats_demanded_angle(1) - slats_demanded_angle).abs() < angle_delta);
+            assert_eq!(self.get_flaps_conf(0), conf);
             assert_eq!(self.get_flaps_conf(1), conf);
         }
     }
-    impl TestBed for A320FlapsTestBed {
-        type Aircraft = A320FlapsTestAircraft;
+    impl TestBed for A380FlapsTestBed {
+        type Aircraft = A380FlapsTestAircraft;
 
-        fn test_bed(&self) -> &SimulationTestBed<A320FlapsTestAircraft> {
+        fn test_bed(&self) -> &SimulationTestBed<A380FlapsTestAircraft> {
             &self.test_bed
         }
 
-        fn test_bed_mut(&mut self) -> &mut SimulationTestBed<A320FlapsTestAircraft> {
+        fn test_bed_mut(&mut self) -> &mut SimulationTestBed<A380FlapsTestAircraft> {
             &mut self.test_bed
         }
     }
 
-    fn test_bed() -> A320FlapsTestBed {
-        A320FlapsTestBed::new()
+    fn test_bed() -> A380FlapsTestBed {
+        A380FlapsTestBed::new()
     }
 
-    fn test_bed_with() -> A320FlapsTestBed {
+    fn test_bed_with() -> A380FlapsTestBed {
         test_bed()
     }
 
@@ -762,22 +696,6 @@ mod tests {
 
         assert!(test_bed.contains_variable_with_name("FLAPS_CONF_INDEX"));
 
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_1"));
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_2"));
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_3"));
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_4"));
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_5"));
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_6"));
-        assert!(test_bed.contains_variable_with_name("SFCC_1_FAP_7"));
-
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_1"));
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_2"));
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_3"));
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_4"));
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_5"));
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_6"));
-        assert!(test_bed.contains_variable_with_name("SFCC_2_FAP_7"));
-
         assert!(test_bed.contains_variable_with_name("SFCC_1_FLAP_ACTUAL_POSITION_WORD"));
         assert!(test_bed.contains_variable_with_name("SFCC_2_FLAP_ACTUAL_POSITION_WORD"));
 
@@ -792,199 +710,12 @@ mod tests {
     }
 
     #[test]
-    fn sfcc_faps() {
-        let mut test_bed = test_bed_with()
-            .set_green_hyd_pressure()
-            .set_yellow_hyd_pressure()
-            .set_blue_hyd_pressure()
-            .set_indicated_airspeed(0.)
-            .set_flaps_handle_position(0)
-            .run_one_tick();
-
-        assert!(!test_bed.get_fap_1(0));
-        assert!(!test_bed.get_fap_2(0));
-        assert!(!test_bed.get_fap_3(0));
-        assert!(test_bed.get_fap_4(0));
-        assert!(!test_bed.get_fap_5(0));
-        assert!(!test_bed.get_fap_6(0));
-        assert!(!test_bed.get_fap_7(0));
-
-        assert!(!test_bed.get_fap_1(1));
-        assert!(!test_bed.get_fap_2(1));
-        assert!(!test_bed.get_fap_3(1));
-        assert!(test_bed.get_fap_4(1));
-        assert!(!test_bed.get_fap_5(1));
-        assert!(!test_bed.get_fap_6(1));
-        assert!(!test_bed.get_fap_7(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(1));
-        assert!(!test_bed.read_sfcc_fap_2_word(1));
-        assert!(!test_bed.read_sfcc_fap_3_word(1));
-        assert!(test_bed.read_sfcc_fap_4_word(1));
-        assert!(!test_bed.read_sfcc_fap_5_word(1));
-        assert!(!test_bed.read_sfcc_fap_6_word(1));
-        assert!(!test_bed.read_sfcc_fap_7_word(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(2));
-        assert!(!test_bed.read_sfcc_fap_2_word(2));
-        assert!(!test_bed.read_sfcc_fap_3_word(2));
-        assert!(test_bed.read_sfcc_fap_4_word(2));
-        assert!(!test_bed.read_sfcc_fap_5_word(2));
-        assert!(!test_bed.read_sfcc_fap_6_word(2));
-        assert!(!test_bed.read_sfcc_fap_7_word(2));
-
-        test_bed = test_bed
-            .set_flaps_handle_position(1)
-            .run_waiting_for(Duration::from_secs(60));
-
-        assert!(!test_bed.get_fap_1(0));
-        assert!(test_bed.get_fap_2(0));
-        assert!(!test_bed.get_fap_3(0));
-        assert!(!test_bed.get_fap_4(0));
-        assert!(!test_bed.get_fap_5(0));
-        assert!(!test_bed.get_fap_6(0));
-        assert!(test_bed.get_fap_7(0));
-
-        assert!(!test_bed.get_fap_1(1));
-        assert!(test_bed.get_fap_2(1));
-        assert!(!test_bed.get_fap_3(1));
-        assert!(!test_bed.get_fap_4(1));
-        assert!(!test_bed.get_fap_5(1));
-        assert!(!test_bed.get_fap_6(1));
-        assert!(test_bed.get_fap_7(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(1));
-        assert!(test_bed.read_sfcc_fap_2_word(1));
-        assert!(!test_bed.read_sfcc_fap_3_word(1));
-        assert!(!test_bed.read_sfcc_fap_4_word(1));
-        assert!(!test_bed.read_sfcc_fap_5_word(1));
-        assert!(!test_bed.read_sfcc_fap_6_word(1));
-        assert!(test_bed.read_sfcc_fap_7_word(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(2));
-        assert!(test_bed.read_sfcc_fap_2_word(2));
-        assert!(!test_bed.read_sfcc_fap_3_word(2));
-        assert!(!test_bed.read_sfcc_fap_4_word(2));
-        assert!(!test_bed.read_sfcc_fap_5_word(2));
-        assert!(!test_bed.read_sfcc_fap_6_word(2));
-        assert!(test_bed.read_sfcc_fap_7_word(2));
-
-        test_bed = test_bed
-            .set_flaps_handle_position(2)
-            .run_waiting_for(Duration::from_secs(60));
-
-        assert!(!test_bed.get_fap_1(0));
-        assert!(test_bed.get_fap_2(0));
-        assert!(!test_bed.get_fap_3(0));
-        assert!(!test_bed.get_fap_4(0));
-        assert!(!test_bed.get_fap_5(0));
-        assert!(!test_bed.get_fap_6(0));
-        assert!(test_bed.get_fap_7(0));
-
-        assert!(!test_bed.get_fap_1(1));
-        assert!(test_bed.get_fap_2(1));
-        assert!(!test_bed.get_fap_3(1));
-        assert!(!test_bed.get_fap_4(1));
-        assert!(!test_bed.get_fap_5(1));
-        assert!(!test_bed.get_fap_6(1));
-        assert!(test_bed.get_fap_7(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(1));
-        assert!(test_bed.read_sfcc_fap_2_word(1));
-        assert!(!test_bed.read_sfcc_fap_3_word(1));
-        assert!(!test_bed.read_sfcc_fap_4_word(1));
-        assert!(!test_bed.read_sfcc_fap_5_word(1));
-        assert!(!test_bed.read_sfcc_fap_6_word(1));
-        assert!(test_bed.read_sfcc_fap_7_word(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(2));
-        assert!(test_bed.read_sfcc_fap_2_word(2));
-        assert!(!test_bed.read_sfcc_fap_3_word(2));
-        assert!(!test_bed.read_sfcc_fap_4_word(2));
-        assert!(!test_bed.read_sfcc_fap_5_word(2));
-        assert!(!test_bed.read_sfcc_fap_6_word(2));
-        assert!(test_bed.read_sfcc_fap_7_word(2));
-
-        test_bed = test_bed
-            .set_flaps_handle_position(3)
-            .run_waiting_for(Duration::from_secs(60));
-
-        assert!(!test_bed.get_fap_1(0));
-        assert!(test_bed.get_fap_2(0));
-        assert!(test_bed.get_fap_3(0));
-        assert!(!test_bed.get_fap_4(0));
-        assert!(test_bed.get_fap_5(0));
-        assert!(!test_bed.get_fap_6(0));
-        assert!(test_bed.get_fap_7(0));
-
-        assert!(!test_bed.get_fap_1(1));
-        assert!(test_bed.get_fap_2(1));
-        assert!(test_bed.get_fap_3(1));
-        assert!(!test_bed.get_fap_4(1));
-        assert!(test_bed.get_fap_5(1));
-        assert!(!test_bed.get_fap_6(1));
-        assert!(test_bed.get_fap_7(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(1));
-        assert!(test_bed.read_sfcc_fap_2_word(1));
-        assert!(test_bed.read_sfcc_fap_3_word(1));
-        assert!(!test_bed.read_sfcc_fap_4_word(1));
-        assert!(test_bed.read_sfcc_fap_5_word(1));
-        assert!(!test_bed.read_sfcc_fap_6_word(1));
-        assert!(test_bed.read_sfcc_fap_7_word(1));
-
-        assert!(!test_bed.read_sfcc_fap_1_word(2));
-        assert!(test_bed.read_sfcc_fap_2_word(2));
-        assert!(test_bed.read_sfcc_fap_3_word(2));
-        assert!(!test_bed.read_sfcc_fap_4_word(2));
-        assert!(test_bed.read_sfcc_fap_5_word(2));
-        assert!(!test_bed.read_sfcc_fap_6_word(2));
-        assert!(test_bed.read_sfcc_fap_7_word(2));
-
-        test_bed = test_bed
-            .set_flaps_handle_position(4)
-            .run_waiting_for(Duration::from_secs(60));
-
-        assert!(test_bed.get_fap_1(0));
-        assert!(test_bed.get_fap_2(0));
-        assert!(test_bed.get_fap_3(0));
-        assert!(!test_bed.get_fap_4(0));
-        assert!(test_bed.get_fap_5(0));
-        assert!(test_bed.get_fap_6(0));
-        assert!(test_bed.get_fap_7(0));
-
-        assert!(test_bed.get_fap_1(1));
-        assert!(test_bed.get_fap_2(1));
-        assert!(test_bed.get_fap_3(1));
-        assert!(!test_bed.get_fap_4(1));
-        assert!(test_bed.get_fap_5(1));
-        assert!(test_bed.get_fap_6(1));
-        assert!(test_bed.get_fap_7(1));
-
-        assert!(test_bed.read_sfcc_fap_1_word(1));
-        assert!(test_bed.read_sfcc_fap_2_word(1));
-        assert!(test_bed.read_sfcc_fap_3_word(1));
-        assert!(!test_bed.read_sfcc_fap_4_word(1));
-        assert!(test_bed.read_sfcc_fap_5_word(1));
-        assert!(test_bed.read_sfcc_fap_6_word(1));
-        assert!(test_bed.read_sfcc_fap_7_word(1));
-
-        assert!(test_bed.read_sfcc_fap_1_word(2));
-        assert!(test_bed.read_sfcc_fap_2_word(2));
-        assert!(test_bed.read_sfcc_fap_3_word(2));
-        assert!(!test_bed.read_sfcc_fap_4_word(2));
-        assert!(test_bed.read_sfcc_fap_5_word(2));
-        assert!(test_bed.read_sfcc_fap_6_word(2));
-        assert!(test_bed.read_sfcc_fap_7_word(2));
-    }
-
-    #[test]
     fn flaps_test_correct_bus_output_clean_config() {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_yellow_hyd_pressure()
             .set_blue_hyd_pressure()
-            .set_indicated_airspeed(0.)
+            .set_indicated_airspeed(160.)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1031,7 +762,7 @@ mod tests {
             .set_green_hyd_pressure()
             .set_yellow_hyd_pressure()
             .set_blue_hyd_pressure()
-            .set_indicated_airspeed(200.)
+            .set_indicated_airspeed(215.)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1049,7 +780,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
@@ -1096,14 +827,14 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(14));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(19));
-        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
+        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(23));
@@ -1113,7 +844,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(14));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(19));
-        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
+        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(23));
@@ -1143,7 +874,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
@@ -1151,7 +882,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
-        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
+        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(23));
 
@@ -1161,7 +892,7 @@ mod tests {
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
-        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
+        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(23));
     }
@@ -1190,26 +921,26 @@ mod tests {
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(31));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
-        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(14));
-        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
+        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(14));
+        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(20));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(21));
-        assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
+        assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(23));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(13));
-        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(14));
-        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
+        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(14));
+        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(15));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(19));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(20));
         assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(21));
-        assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
+        assert!(test_bed.read_slat_flap_actual_position_word(2).get_bit(22));
         assert!(!test_bed.read_slat_flap_actual_position_word(2).get_bit(23));
     }
 
@@ -1237,7 +968,7 @@ mod tests {
         assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(21));
         assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(26));
 
-        test_bed = test_bed.run_waiting_for(Duration::from_secs(45));
+        test_bed = test_bed.run_waiting_for(Duration::from_secs(60));
 
         assert!(!test_bed.read_slat_flap_actual_position_word(1).get_bit(12));
         assert!(test_bed.read_slat_flap_actual_position_word(1).get_bit(13));
@@ -1262,65 +993,65 @@ mod tests {
 
     // Tests flaps configuration and angles for regular
     // increasing handle transitions, i.e 0->1->2->3->4 in sequence
-    // below 100 knots
+    // below 205 knots
     #[test]
-    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_below_100() {
+    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_below_205() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(50.)
+            .set_indicated_airspeed(160.)
             .run_one_tick();
 
         test_bed.test_flap_conf(0, 0., 0., FlapsConf::Conf0, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 120.22, 222.27, FlapsConf::Conf1F, angle_delta);
+        test_bed.test_flap_conf(1, 108.28, 247.27, FlapsConf::Conf1F, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 218.91, 284.65, FlapsConf::ConfFull, angle_delta);
     }
 
     // Tests flaps configuration and angles for regular
     // increasing handle transitions, i.e 0->1->2->3->4 in sequence
-    // above 100 knots
+    // above 205 knots
     #[test]
-    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_above_100() {
+    fn flaps_test_regular_handle_increase_transitions_flaps_target_airspeed_above_205() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(150.)
+            .set_indicated_airspeed(215.)
             .run_one_tick();
 
         test_bed.test_flap_conf(0, 0., 0., FlapsConf::Conf0, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 0., 222.27, FlapsConf::Conf1, angle_delta);
+        test_bed.test_flap_conf(1, 0., 247.27, FlapsConf::Conf1, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 154.65, 284.65, FlapsConf::Conf2S, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
     }
 
-    //Tests regular transition 2->1 below and above 210 knots
+    //Tests regular transition 2->1 below and above 212 knots
     #[test]
     fn flaps_test_regular_handle_transition_pos_2_to_1() {
         let mut test_bed = test_bed_with()
@@ -1350,7 +1081,7 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1);
     }
 
-    //Tests transition between Conf1F to Conf1 above 210 knots
+    //Tests transition between Conf1F to Conf1 above 212 knots
     #[test]
     fn flaps_test_regular_handle_transition_pos_1_to_1() {
         let mut test_bed = test_bed_with()
@@ -1375,30 +1106,30 @@ mod tests {
 
     // Tests flaps configuration and angles for regular
     // decreasing handle transitions, i.e 4->3->2->1->0 in sequence
-    // below 210 knots
+    // below 212 knots
     #[test]
-    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_below_210() {
+    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_below_212() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(150.)
+            .set_indicated_airspeed(160.)
             .run_one_tick();
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 218.91, 284.65, FlapsConf::ConfFull, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 120.22, 222.27, FlapsConf::Conf1F, angle_delta);
+        test_bed.test_flap_conf(1, 108.28, 247.27, FlapsConf::Conf1F, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
 
@@ -1409,7 +1140,7 @@ mod tests {
     // decreasing handle transitions, i.e 4->3->2->1->0 in sequence
     // above 210 knots
     #[test]
-    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_above_210() {
+    fn flaps_test_regular_decrease_handle_transition_flaps_target_airspeed_above_212() {
         let angle_delta: f64 = 0.1;
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
@@ -1418,19 +1149,19 @@ mod tests {
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
 
-        test_bed.test_flap_conf(4, 251.97, 334.16, FlapsConf::ConfFull, angle_delta);
+        test_bed.test_flap_conf(4, 218.91, 284.65, FlapsConf::ConfFull, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
 
-        test_bed.test_flap_conf(3, 168.35, 272.27, FlapsConf::Conf3, angle_delta);
+        test_bed.test_flap_conf(3, 194.03, 284.65, FlapsConf::Conf3, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
 
-        test_bed.test_flap_conf(2, 145.51, 272.27, FlapsConf::Conf2, angle_delta);
+        test_bed.test_flap_conf(2, 154.65, 247.27, FlapsConf::Conf2, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(1).run_one_tick();
 
-        test_bed.test_flap_conf(1, 0., 222.27, FlapsConf::Conf1, angle_delta);
+        test_bed.test_flap_conf(1, 0., 247.27, FlapsConf::Conf1, angle_delta);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
 
@@ -1455,23 +1186,23 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf2);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf3);
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf3);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::ConfFull);
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::ConfFull);
 
         test_bed = test_bed
-            .set_indicated_airspeed(110.)
+            .set_indicated_airspeed(160.)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1546,7 +1277,7 @@ mod tests {
 
         test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(110.)
+            .set_indicated_airspeed(210.)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1563,7 +1294,7 @@ mod tests {
 
         test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(110.)
+            .set_indicated_airspeed(205.)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1642,8 +1373,8 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf2);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
 
         test_bed = test_bed.set_flaps_handle_position(2).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf2);
@@ -1654,7 +1385,7 @@ mod tests {
     fn flaps_test_irregular_handle_transition_init_pos_3() {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(150.)
+            .set_indicated_airspeed(205.)
             .set_flaps_handle_position(3)
             .run_one_tick();
 
@@ -1696,8 +1427,8 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf3);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
 
         test_bed = test_bed.set_flaps_handle_position(3).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf3);
@@ -1708,7 +1439,7 @@ mod tests {
     fn flaps_test_irregular_handle_transition_init_pos_4() {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
-            .set_indicated_airspeed(150.)
+            .set_indicated_airspeed(205.)
             .set_flaps_handle_position(4)
             .run_one_tick();
 
@@ -1750,8 +1481,8 @@ mod tests {
         assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::ConfFull);
 
         test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
-        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
-        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F); // alpha lock
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F); // alpha lock
 
         test_bed = test_bed.set_flaps_handle_position(4).run_one_tick();
         assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::ConfFull);
@@ -1772,6 +1503,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1798,6 +1530,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1806,7 +1539,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(2)
-            .run_waiting_for(Duration::from_secs(20));
+            .run_waiting_for(Duration::from_secs(35));
 
         assert!(
             (test_bed.get_flaps_fppu_feedback() - test_bed.get_flaps_demanded_angle(0)).abs()
@@ -1824,6 +1557,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(2)
             .run_one_tick();
 
@@ -1850,6 +1584,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(3)
             .run_waiting_for(Duration::from_secs(20));
 
@@ -1876,6 +1611,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(0)
             .run_one_tick();
 
@@ -1884,7 +1620,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(1)
-            .run_waiting_for(Duration::from_secs(30));
+            .run_waiting_for(Duration::from_secs(60));
 
         assert!(
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(0)).abs()
@@ -1911,7 +1647,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(1)
-            .run_waiting_for(Duration::from_secs(30));
+            .run_waiting_for(Duration::from_secs(60));
 
         assert!(
             (test_bed.get_flaps_fppu_feedback() - test_bed.get_flaps_demanded_angle(0)).abs()
@@ -1937,6 +1673,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(1)
             .run_one_tick();
 
@@ -1945,7 +1682,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(2)
-            .run_waiting_for(Duration::from_secs(40));
+            .run_waiting_for(Duration::from_secs(60));
 
         assert!(
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(0)).abs()
@@ -1963,6 +1700,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(2)
             .run_one_tick();
 
@@ -1971,7 +1709,7 @@ mod tests {
 
         test_bed = test_bed
             .set_flaps_handle_position(3)
-            .run_waiting_for(Duration::from_secs(40));
+            .run_waiting_for(Duration::from_secs(60));
 
         assert!(
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(0)).abs()
@@ -1989,6 +1727,7 @@ mod tests {
         let mut test_bed = test_bed_with()
             .set_green_hyd_pressure()
             .set_indicated_airspeed(0.)
+            .set_on_ground(true)
             .set_flaps_handle_position(3)
             .run_one_tick();
 
@@ -2007,5 +1746,134 @@ mod tests {
             (test_bed.get_slats_fppu_feedback() - test_bed.get_slats_demanded_angle(1)).abs()
                 <= angle_delta
         );
+    }
+
+    #[test]
+    fn config_test_flrs_full() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_indicated_airspeed(186.)
+            .set_flaps_handle_position(0)
+            .run_one_tick();
+
+        test_bed = test_bed
+            .set_flaps_handle_position(4)
+            .run_waiting_for(Duration::from_secs(50));
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf3);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf3);
+    }
+
+    #[test]
+    fn config_test_flrs_3() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_indicated_airspeed(200.)
+            .set_flaps_handle_position(0)
+            .run_one_tick();
+
+        test_bed = test_bed
+            .set_flaps_handle_position(3)
+            .run_waiting_for(Duration::from_secs(50));
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf2S);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf2S);
+    }
+
+    #[test]
+    fn config_test_flrs_2() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_indicated_airspeed(224.)
+            .set_flaps_handle_position(0)
+            .run_one_tick();
+
+        test_bed = test_bed
+            .set_flaps_handle_position(2)
+            .run_waiting_for(Duration::from_secs(50));
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
+    }
+
+    #[test]
+    fn config_test_cruise_baulk() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_indicated_airspeed(280.)
+            .set_on_ground(false)
+            .set_flaps_handle_position(0)
+            .run_one_tick();
+
+        test_bed = test_bed
+            .set_flaps_handle_position(1)
+            .run_waiting_for(Duration::from_secs(20));
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert!(test_bed.read_slat_flap_system_status_word(1).get_bit(25));
+        assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(25));
+    }
+
+    #[test]
+    fn config_test_alpha_lock_speed() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_indicated_airspeed(130.)
+            .set_angle_of_attack(Angle::new::<degree>(5.))
+            .set_on_ground(false)
+            .set_flaps_handle_position(1)
+            .run_one_tick();
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
+        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
+        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
+
+        test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
+        assert!(test_bed.read_slat_flap_system_status_word(1).get_bit(24));
+        assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(24));
+
+        test_bed = test_bed.set_indicated_airspeed(200.).run_one_tick();
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
+        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
+    }
+
+    #[test]
+    fn config_test_alpha_lock_aoa() {
+        let mut test_bed = test_bed_with()
+            .set_green_hyd_pressure()
+            .set_indicated_airspeed(200.)
+            .set_angle_of_attack(Angle::new::<degree>(10.))
+            .set_on_ground(false)
+            .set_flaps_handle_position(1)
+            .run_one_tick();
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
+        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
+        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
+
+        test_bed = test_bed.set_flaps_handle_position(0).run_one_tick();
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf1F);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf1F);
+        assert!(test_bed.read_slat_flap_system_status_word(1).get_bit(24));
+        assert!(test_bed.read_slat_flap_system_status_word(2).get_bit(24));
+
+        test_bed = test_bed
+            .set_angle_of_attack(Angle::new::<degree>(9.1))
+            .run_one_tick();
+
+        assert_eq!(test_bed.get_flaps_conf(0), FlapsConf::Conf0);
+        assert_eq!(test_bed.get_flaps_conf(1), FlapsConf::Conf0);
+        assert!(!test_bed.read_slat_flap_system_status_word(1).get_bit(24));
+        assert!(!test_bed.read_slat_flap_system_status_word(2).get_bit(24));
     }
 }
